// src/lib/api/models/TicketPublic.ts
export interface TicketPublic {
    /**
     * 
     * @type {string}
     * @memberof TicketPublic
     */
    subject: string;
    /**
     * 
     * @type {string}
     * @memberof TicketPublic
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof TicketPublic
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof TicketPublic
     */
    priority?: string;
    /**
     * 
     * @type {string}
     * @memberof TicketPublic
     */
    category?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketPublic
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof TicketPublic
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof TicketPublic
     */
    updatedAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof TicketPublic
     */
    resolvedAt: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof TicketPublic
     */
    closedAt: Date | null;
    /**
     * 
     * @type {string}
     * @memberof TicketPublic
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof TicketPublic
     */
    assignedTo: string | null;
    /**
     * 
     * @type {number}
     * @memberof TicketPublic
     */
    responseCount?: number;
}