// src/lib/api/models/TicketsPublic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TicketPublic } from './TicketPublic';
import {
    TicketPublicFromJSON,
    TicketPublicFromJSONTyped,
    TicketPublicToJSON,
    TicketPublicToJSONTyped,
} from './TicketPublic';

/**
 * 
 * @export
 * @interface TicketsPublic
 */
export interface TicketsPublic {
    /**
     * 
     * @type {Array<TicketPublic>}
     * @memberof TicketsPublic
     */
    data: Array<TicketPublic>;
    /**
     * 
     * @type {number}
     * @memberof TicketsPublic
     */
    count: number;
}

/**
 * Check if a given object implements the TicketsPublic interface.
 */
export function instanceOfTicketsPublic(value: object): value is TicketsPublic {
    if (!('data' in value) || value['data'] === undefined) return false;
    if (!('count' in value) || value['count'] === undefined) return false;
    return true;
}

export function TicketsPublicFromJSON(json: any): TicketsPublic {
    return TicketsPublicFromJSONTyped(json, false);
}

export function TicketsPublicFromJSONTyped(json: any, ignoreDiscriminator: boolean): TicketsPublic {
    if (json == null) {
        return json;
    }
    return {
        
        'data': ((json['data'] as Array<any>).map(TicketPublicFromJSON)),
        'count': json['count'],
    };
}

export function TicketsPublicToJSON(json: any): TicketsPublic {
    return TicketsPublicToJSONTyped(json, false);
}

export function TicketsPublicToJSONTyped(value?: TicketsPublic | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': ((value['data'] as Array<any>).map(TicketPublicToJSON)),
        'count': value['count'],
    };
}