// src/lib/api/models/TicketCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TicketCreate
 */
export interface TicketCreate {
    /**
     * 
     * @type {string}
     * @memberof TicketCreate
     */
    subject: string;
    /**
     * 
     * @type {string}
     * @memberof TicketCreate
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof TicketCreate
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof TicketCreate
     */
    priority?: string;
    /**
     * 
     * @type {string}
     * @memberof TicketCreate
     */
    category?: string | null;
}

/**
 * Check if a given object implements the TicketCreate interface.
 */
export function instanceOfTicketCreate(value: object): value is TicketCreate {
    if (!('subject' in value) || value['subject'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    return true;
}

export function TicketCreateFromJSON(json: any): TicketCreate {
    return TicketCreateFromJSONTyped(json, false);
}

export function TicketCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TicketCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'subject': json['subject'],
        'description': json['description'],
        'status': json['status'] == null ? undefined : json['status'],
        'priority': json['priority'] == null ? undefined : json['priority'],
        'category': json['category'] == null ? undefined : json['category'],
    };
}

export function TicketCreateToJSON(json: any): TicketCreate {
    return TicketCreateToJSONTyped(json, false);
}

export function TicketCreateToJSONTyped(value?: TicketCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'subject': value['subject'],
        'description': value['description'],
        'status': value['status'],
        'priority': value['priority'],
        'category': value['category'],
    };
}