// src/lib/api/models/TicketUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TicketUpdate
 */
export interface TicketUpdate {
    /**
     * 
     * @type {string}
     * @memberof TicketUpdate
     */
    subject?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketUpdate
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketUpdate
     */
    status?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketUpdate
     */
    priority?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketUpdate
     */
    category?: string | null;
}

/**
 * Check if a given object implements the TicketUpdate interface.
 */
export function instanceOfTicketUpdate(value: object): value is TicketUpdate {
    return true;
}

export function TicketUpdateFromJSON(json: any): TicketUpdate {
    return TicketUpdateFromJSONTyped(json, false);
}

export function TicketUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TicketUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'subject': json['subject'] == null ? undefined : json['subject'],
        'description': json['description'] == null ? undefined : json['description'],
        'status': json['status'] == null ? undefined : json['status'],
        'priority': json['priority'] == null ? undefined : json['priority'],
        'category': json['category'] == null ? undefined : json['category'],
    };
}

export function TicketUpdateToJSON(json: any): TicketUpdate {
    return TicketUpdateToJSONTyped(json, false);
}

export function TicketUpdateToJSONTyped(value?: TicketUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'subject': value['subject'],
        'description': value['description'],
        'status': value['status'],
        'priority': value['priority'],
        'category': value['category'],
    };
}